requirements.txt
greenlet==3.2.4
markdown-it-py==4.0.0
mdurl==0.1.2
networkx==3.6
psycopg2-binary==2.9.11
Pygments==2.19.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
rich==14.2.0
setuptools==80.9.0
six==1.17.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
wheel==0.45.1
